+++ esp-wifi/src/compat/malloc.rs
@@ -1,20 +1,16 @@
 use core::alloc::Layout;
 
-use crate::HEAP;
-
 #[no_mangle]
 pub unsafe extern "C" fn malloc(size: usize) -> *mut u8 {
     trace!("alloc {}", size);
 
     let total_size = size + 4;
 
-    let layout = Layout::from_size_align_unchecked(total_size, 4);
-    let ptr = critical_section::with(|cs| {
-        HEAP.borrow_ref_mut(cs)
-            .allocate_first_fit(layout)
-            .ok()
-            .map_or(core::ptr::null_mut(), |allocation| allocation.as_ptr())
-    });
+    extern "C" {
+        fn esp_wifi_allocate_from_internal_ram(size: usize) -> *mut u8;
+    }
+
+    let ptr = unsafe { esp_wifi_allocate_from_internal_ram(total_size) };
 
     if ptr.is_null() {
         warn!("Unable to allocate {} bytes", size);
@@ -37,10 +33,7 @@ pub unsafe extern "C" fn free(ptr: *mut u8) {
     let total_size = *(ptr as *const usize);
 
     let layout = Layout::from_size_align_unchecked(total_size, 4);
-    critical_section::with(|cs| {
-        HEAP.borrow_ref_mut(cs)
-            .deallocate(core::ptr::NonNull::new_unchecked(ptr), layout)
-    });
+    alloc::alloc::dealloc(ptr, layout);
 }
 
 #[no_mangle]
@@ -78,5 +71,24 @@ unsafe extern "C" fn realloc(ptr: *mut u8, new_size: usize) -> *mut u8 {
             free(ptr);
         }
         p
+    }
+}
+
+#[cfg(feature = "esp-alloc")]
+#[doc(hidden)]
+#[no_mangle]
+pub extern "C" fn esp_wifi_free_internal_heap() -> usize {
+    esp_alloc::HEAP.free_caps(esp_alloc::MemoryCapability::Internal.into())
+}
+
+#[cfg(feature = "esp-alloc")]
+#[doc(hidden)]
+#[no_mangle]
+pub extern "C" fn esp_wifi_allocate_from_internal_ram(size: usize) -> *mut u8 {
+    unsafe {
+        esp_alloc::HEAP.alloc_caps(
+            esp_alloc::MemoryCapability::Internal.into(),
+            core::alloc::Layout::from_size_align_unchecked(size, 4),
+        )
     }
 }