+++ examples/src/bin/rtc_time.rs
@@ -5,6 +5,8 @@
 #![no_std]
 #![no_main]
 
+use core::time::Duration;
+
 use esp_backtrace as _;
 use esp_hal::{delay::Delay, prelude::*, rtc_cntl::Rtc};
 
@@ -16,7 +18,18 @@ fn main() -> ! {
     let delay = Delay::new();
 
     loop {
-        esp_println::println!("rtc time in milliseconds is {}", rtc.get_time_ms());
+        esp_println::println!(
+            "rtc time in milliseconds is {}",
+            rtc.current_time().and_utc().timestamp_millis()
+        );
         delay.delay_millis(1000);
+
+        // Set the time to half a second in the past
+        let new_time = rtc.current_time() - Duration::from_millis(500);
+        esp_println::println!(
+            "setting rtc time to {} milliseconds",
+            new_time.and_utc().timestamp_millis()
+        );
+        rtc.set_current_time(new_time);
     }
 }